// apps/api/src/common/enums/permission.enum.ts
export enum Permission {
  READ_USER = "READ_USER",
  WRITE_USER_ROLE = "WRITE_USER_ROLE",
  READ_PROJECT = "READ_PROJECT",
  WRITE_PROJECT = "WRITE_PROJECT",
  READ_RESOURCE = "READ_RESOURCE",
  WRITE_RESOURCE = "WRITE_RESOURCE",
  READ_GITLAB = "READ_GITLAB",
  WRITE_GITLAB = "WRITE_GITLAB",
  READ_LLM = "READ_LLM",
  WRITE_LLM = "WRITE_LLM",
  READ_VECTORDB = "READ_VECTORDB",
  WRITE_VECTORDB = "WRITE_VECTORDB",
  READ_AUDIT_LOG = "READ_AUDIT_LOG",
  READ_ACCESS_LOG = "READ_ACCESS_LOG",
}